[Refactoring] Move MemberwiseInitLocalRefactoring to its own file
