+++ lib/Refactoring/CMakeLists.txt
@@ -13,6 +13,7 @@ add_swift_host_library(swiftRefactoring STATIC
   ExtractFunction.cpp
   FillProtocolStubs.cpp
   LocalizeString.cpp
+  MemberwiseInitLocalRefactoring.cpp
   MoveMembersToExtension.cpp
   Refactoring.cpp
   Renamer.cpp
